# frontend/public/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Cassandra API Reference",
    "version": "1.0.0",
    "description": "Documentación viva de la API autogenerada a partir de las colecciones de Bruno."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Servidor Local"
    },
    {
      "url": "/",
      "description": "Servidor Actual"
    }
  ],
...
  "paths": {
    "/api/proyects/{proyect_id}/actividad": {
      "get": {
        "tags": [
          "Actividad"
        ],
        "summary": "Actividad de un proyecto",
        "operationId": "get__api_proyects__proyect_id__actividad",
        "parameters": [
          {
            "name": "proyect_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "ID numérico de proyect"
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "401": {
            "description": "No autorizado / Token inválido"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/actividad/resumen": {
      "get": {
        "tags": [
          "Actividad"
        ],
        "summary": "Resumen de actividad",
        "operationId": "get__api_actividad_resumen",
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "401": {
            "description": "No autorizado / Token inválido"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "login",
        "operationId": "post__api_auth_login",
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "401": {
            "description": "No autorizado / Token inválido"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "logout",
        "operationId": "post__api_auth_logout",
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "401": {
            "description": "No autorizado / Token inválido"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "refresh",
        "operationId": "post__api_auth_refresh",
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "401": {
            "description": "No autorizado / Token inválido"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/api/proyects/{proyect_id}/documentos": {
      "post": {
        "tags": [
          "Documentos proyectos"
        ],
        "summary": "Create documento",
        "operationId": "post__api_proyects__proyect_id__documentos",
        "parameters": [
          {
            "name": "proyect_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "ID numérico de proyect"
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "401": {
            "description": "No autorizado / Token inválido"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "tags": [
          "Documentos proyectos"
        ],
        "summary": "Get all by proyecto",
        "operationId": "get__api_proyects__proyect_id__documentos",
        "parameters": [
          {
            "name": "proyect_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "ID numérico de proyect"
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "401": {
            "description": "No autorizado / Token inválido"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/documentos/{documento_id}": {
      "delete": {
        "tags": [
          "Documentos proyectos"
        ],
        "summary": "Delete documento",
        "operationId": "delete__api_documentos__documento_id_",
        "parameters": [
          {
            "name": "documento_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "ID numérico de documento"
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "401": {
            "description": "No autorizado / Token inválido"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "tags": [
          "Documentos proyectos"
        ],
        "summary": "Get by ID",
        "operationId": "get__api_documentos__documento_id_",
        "parameters": [
          {
            "name": "documento_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "ID numérico de documento"
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "401": {
            "description": "No autorizado / Token inválido"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Documentos proyectos"
        ],
        "summary": "Update documento",
        "operationId": "put__api_documentos__documento_id_",
        "parameters": [
          {
            "name": "documento_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "ID numérico de documento"
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "401": {
            "description": "No autorizado / Token inválido"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/finanzas/bancos/{banco_id}": {
      "put": {
        "tags": [
          "Finanzas"
        ],
        "summary": "Actualizar Banco",
        "operationId": "put__api_finanzas_bancos__banco_id_",
        "parameters": [
          {
            "name": "banco_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "ID numérico de banco"
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "401": {
            "description": "No autorizado / Token inválido"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Finanzas"
        ],
        "summary": "Eliminar Banco",
        "operationId": "delete__api_finanzas_bancos__banco_id_",
        "parameters": [
          {
            "name": "banco_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "ID numérico de banco"
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "401": {
            "description": "No autorizado / Token inválido"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/finanzas/grupos/{grupo_id}": {
      "put": {
        "tags": [
          "Finanzas"
        ],
        "summary": "Actualizar Grupo Finanzas",
        "operationId": "put__api_finanzas_grupos__grupo_id_",
        "parameters": [
          {
            "name": "grupo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "ID numérico de grupo"
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "401": {
            "description": "No autorizado / Token inválido"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Finanzas"
        ],
        "summary": "Eliminar Grupo Finanzas",
        "operationId": "delete__api_finanzas_grupos__grupo_id_",
        "parameters": [
          {
            "name": "grupo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "ID numérico de grupo"
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "401": {
            "description": "No autorizado / Token inválido"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/finanzas/plantilla/{plantilla_id}": {
      "put": {
        "tags": [
          "Finanzas"
        ],
        "summary": "Actualizar Item Plantilla",
        "operationId": "put__api_finanzas_plantilla__plantilla_id_",
        "parameters": [
          {
            "name": "plantilla_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "ID numérico de plantilla"
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "401": {
            "description": "No autorizado / Token inválido"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Finanzas"
        ],
        "summary": "Eliminar Item Plantilla",
        "operationId": "delete__api_finanzas_plantilla__plantilla_id_",
        "parameters": [
          {
            "name": "plantilla_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "ID numérico de plantilla"
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "401": {
            "description": "No autorizado / Token inválido"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "tags": [
          "Finanzas"
        ],
        "summary": "Obtener Item Plantilla por ID",
        "operationId": "get__api_finanzas_plantilla__plantilla_id_",
        "parameters": [
          {
            "name": "plantilla_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "ID numérico de plantilla"
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "401": {
            "description": "No autorizado / Token inválido"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/finanzas/movimientos-esperados/{movimientos-esperado_id}": {
      "put": {
        "tags": [
          "Finanzas"
        ],
        "summary": "Actualizar Movimiento Esperado",
        "operationId": "put__api_finanzas_movimientos_esperados__movimientos_esperado_id_",
        "parameters": [
          {
            "name": "movimientos-esperado_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "ID numérico de movimientos-esperado"
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "401": {
            "description": "No autorizado / Token inválido"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Finanzas"
        ],
        "summary": "Eliminar Movimiento Esperado",
        "operationId": "delete__api_finanzas_movimientos_esperados__movimientos_esperado_id_",
        "parameters": [
          {
            "name": "movimientos-esperado_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "ID numérico de movimientos-esperado"
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "401": {
            "description": "No autorizado / Token inválido"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/finanzas/clonar": {
      "post": {
        "tags": [
          "Finanzas"
        ],
        "summary": "Clonar Periodo",
        "operationId": "post__api_finanzas_clonar",
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "401": {
            "description": "No autorizado / Token inválido"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT de autenticación"
      }
    }
  }
}